requirements.txt
Flask==0.10.1
Flask-SocketIO==0.4.2
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
geopy==1.6.1
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.3
gunicorn==0.16.1
itsdangerous==0.24
oauthlib==0.7.2
requests==2.4.3
requests-oauthlib==0.4.1
six==1.7.3
tweepy==3.1.0
wsgiref==0.1.2
